+++ include/ft_print.h
@@ -6,7 +6,7 @@
 /*   By: lgoddijn <lgoddijn@student.codam.nl >      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/31 17:04:23 by lgoddijn          #+#    #+#             */
-/*   Updated: 2024/01/11 15:39:56 by lgoddijn         ###   ########.fr       */
+/*   Updated: 2024/04/14 19:23:33 by lgoddijn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,13 @@
 # include "ft_string.h"
 # include "ft_unistd.h"
 
-int		ft_print_char(const int c);
-int		ft_print_str(const char *buffer);
-int		ft_print_int(int n);
-int		ft_print_uint(const unsigned int n);
-int		ft_print_hex(long long n, const bool is_upper, const bool is_ptr);
-int		ft_print_ptr(const void *ptr_addr);
-int		ft_printf(const char *fmt, ...);
+int32_t	ft_print_char(const int32_t c);
+int32_t	ft_print_str(const char *buffer);
+int32_t	ft_print_int(int32_t n);
+int32_t	ft_print_uint(const uint32_t n);
+int32_t	ft_print_hex(int64_t n, const bool is_upper, const bool is_ptr);
+int32_t	ft_print_ptr(const void *ptr_addr);
+int32_t	ft_printf(const char *fmt, ...);
 void	ft_raise(const char *msg, ...);
 
 #endif
